requirements.txt
streamlit>=1.15.0,<1.30.0
openai==0.28
python-dotenv>=0.21.0,<1.0.0
